Applied modularization.
Pcm2JavaCommon now contains the previous pojo content, and the pojo, ejb
and depInject reactions import it and apply their changes.

The Common projects needs to export the generated code.

Moved the src-gen folder of the individual projects above the
neededPlugins-entry in the classpath, so their locally generated
ChangePropagationSpecification gets preferred over the onr of the
commons project.