Updated to removal of environment ChangePropagationSpecifications.
* Updated generated code: All segments now generate an own
ChangePropagationSpecification. Executors are package-private now.
* Added new composite reactions segments in some projects that
previously relied on the environment ChangePropagationSpecifications to
execute the reactions of all segments contained in the project. Now the
executed segments get explicitly specified by creating a segment which
imports (and therefore combines) the segments one wishes to execute.
  * Java2Pcm.reactions
  * Java2Uml.reactions
  * UmlToJava.reactions
* Reverted some classpath changes from a previous commit in the pcm2java
projects: With the removal of the
environment-ChangePropagationSpecifications this is no longer required,
because the naming conflicts which result when depending on an other
project with reactions segments of the same metamodel-pair ('common') no
longer exists.
* The used ChangePropagationSpecifications in the application projects
were updated to now use the ChangePropagationSpecifications of a
specific segment.
  * EjbJava2PcmChangePropagationSpecification
  * Java2PcmChangePropagationSpecification
  * Pcm2EjbJavaChangePropagationSpecification
  * Pcm2JavaChangePropagationSpecification
  * Pcm2JavaDepInjectChangePropagationSpecification
  * JavaToUmlChangePropagationSpecification
  * UmlToJavaChangePropagationSpecification
  * UmlClass2UmlCompChangePropagation
  * UmlComp2UmlClassChangePropagation
  * PcmToUmlComponentsChangePropagationSpecification
  * UmlToPcmComponentsChangePropagationSpecification
* Updated the Manifests of some projects: Some projects now additionally
depend on the dslruntime.reactions project: Because the new
per-segment-ChangePropagationSpecifications inherit from
AbstractReactionsChangePropagationSpecification which is located in the
dslruntime-project, and those projects can otherwise not figure out that
the new ChangePropagationSpecifications actually (indirectly) implement
ChangePropagationSpecification. (An alternative would be to either not
use an AbstractReactionsChangePropagationSpecification (it's not
strictly required currently), or additionally declare 'implements
ChangePropagationSpecification' in the concrete
ChangePropagationSpecification class)
  * comp2class.tests
  * class2comp.tests
  * pojo.pcm2java.tests
  * pojo.java2pcm.tests
  * ejb.java2pcm.tests
  * umljava
  * pcmumlcomponents
  * pcmjava.ejbtransformations
* Also: The pcm2depInject-project no longer exports the package
'mir.reactions.reactionsPcmToJava.pcm2depInjectJava'. This doesn't seem
to be required anywhere currently..
* Also: Removed the AbstractChangePropagationSpecificationJavaToPcm
class in the src-gen folder of the ejbJava2pcm project: This class
doesn't seem to be used anywhere anymore.