Added test framework and various test utilities.
We perform pairwise tests of all domains. Each test case consists of setting up a model in one source domain and checking if the corresponding model has been created in one other target domain. In order to avoid code duplication when writing these tests for all the different pairwise combinations of source and target domains, the test cases delegate the setup of the source model and the checking of the target model to two domain and test case specific implementations of the 'DomainModel' interface. Tests for different domain combinations simply need to use different combinations of these interface implementations then.

In many cases the models constructed by a domain for its source and target role for a certain test case are the same. Furthermore, the way of how source models are created and target models are checked is often independent of any specific test case and can be described by domain specific implementations of the 'DomainModelTester' interface. It is therefore often possible to further simplify the implementation of the test case specific 'DomainModel' interfaces by using the 'SimpleDomainModel' class.

Test cases for the same concept are grouped in one class. For each of these groups of test cases there is one 'DomainModels' interface that needs to be implemented for each domain and which declares the constants shared by those test cases and acts as provider for their DomainModel implementations. To simplify the implementation of these 'DomainModels' interfaces, there exists a general 'DomainTestModelsBase' base class and domain specific subclasses of it.

Other related changes:
* Added CBSExecutionTestCompiler which sets up the runtime test project and compiles the commonalities.
* Added CBSCommonalitiesExecutionTest which acts as base class for all CBS commonality tests. This invokes the CBSExecutionTestCompiler.
* Added a general FilePathUtil and the domain specific variants JavaFilePathHelper, UmlFilePathHelper and PcmFilePathHelper. These are used by the JavaTestHelper, UmlTestHelper and PcmTestHelper classes, which are themselves used to implement the domain specific DomainModelTesters.
* Added VitruvApplicationTestAdapter interface. This is implemented inside CBSCommonalitiesExecutionTest and allows its subclasses to provide helper objects access to internal methods that would otherwise be restricted to the scope of the test class itself.