Increment 2: Interface and class stubs for empty BasicComponents and component interfaces.
* Added commonalities: CBS:Component, CBS:ComponentInterface, OO:Class, OO:Interface.
  * CBS:Repository contains CBS:Components and CBS:ComponentInterfaces.
  * OO:Package contains OO:Classes and OO:Interfaces.
* Added operators:
  * firstUpper
  * prefix
  * javaPackageCompilationUnits
  * javaCompilationUnitName
  * javaVisibility
  * umlVisibility

A CBS:Component gets mapped to an OO:Package and a contained OO:Class within the repository package. A CBS:Interface gets mapped to an OO:Interface within the contracts package.

The javaPackageCompilationUnits operator also internally updates the resource path of CompilationUnits that are inserted into a package (similar to how it is done for the insertion of sub-packages).

Common Java and UML visibilities are represented by an added Visibility enum. The Java and UML visibility operators map to and from values of this Visibility enum. Since Java elements always have some visibility, we decided to map the absence of a visibility in the intermediate model to Java's public visibility by default.
The intermediate model stores the visibilities as Strings, using the names of the values of the newly added Visibility enum.