Fixed: Some comparisons of PCM models in test cases failed.
The issue was that we manually added a primitive data type for LONG which is missing in PCM's default primitive types repository.
When a PCM model which contains references to at least one of these primitive data types is loaded, the default primitive types repository is loaded as well and will not contain our manually added LONG primitive data type. Due to the container references of each primitive data type to its container repository, the model comparison compares these primitive data type repositories as well and will then fail due to the LONG data type missing in one of them.