Increment 6: Implemented ProvidedRoles.
This is currently implemented as attribute and attribute mappings, because the interface implementations in the OO metamodel are implemented as attribute as well (due to Java using TypeReferences for Java:Class.implements).

A new PCM provided roles attribute mapping operator internally creates the corresponding OperationProvidedRoles for the referenced interfaces. However, this is problematic, because we do not react to changes of the provided interfaces currently (reference OperationProvidedRole.providedInterface__OperationProvidedRole).

Operator changes:
* The PCM type reference operator operates on arbitrary EObjects now (instead of just DataType). This is used when mapping between references to non-DataType PCM objects such as OperationInterface and corresponding intermediate type references.
* The previous use of the PCM type reference operator has been replaced by a new 'pcmDataTypeReference' operator. This operator delegates to the general purpose PCM type reference operator, but only maps to and from PCM's DataType.
* Added OO type references operator, which is similar to the already existing OO type reference operator, but maps between multi-valued type references.
* Added PCM provided roles operator which maps between intermediate interface references and corresponding PCM OperationProvidedRoles. See above.