Classes are only matched as component if they have the suffix 'Impl'.
Added 'hasSuffix' condition operator.

This operator can only be checked, not enforced, because it is likely that the attribute value which the operator operates on has not been set yet at the time participation conditions are enforced. And participation conditions cannot access commonality attributes. This is better suited to happen as part of an attribute mapping instead.

This also avoids accidentally matching classes as components inside the repository's contracts and datatypes packages, because we assume that these packages do not contain classes ending with the 'Impl' suffix.